Fixed small error in the AC method.

+++ src/entities/Project.java
@@ -100,7 +100,7 @@ public class Project {
         }
         return totalPV;
     }
-    public long calculateDuration(){
+    public double calculateDuration(){
         return ChronoUnit.DAYS.between(startDate,endDate);
     }
